Move extractScalars to extractors folder
